1)Refactoring done 2)Exception handlers added 3)Site's Equals  and GetHashCode methods rewritten 4)Unit tests for Site class added
