+++ WebsiteChecker.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 16.0.28922.388
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebsiteChecker", "WebsiteChecker.csproj", "{101D9C9C-23C5-40AA-A927-CCAF2E645FA1}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SiteClassTests", "..\SiteClassTests\SiteClassTests.csproj", "{B6F86E78-30C8-488F-AE8E-FFEC90FB3A13}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{101D9C9C-23C5-40AA-A927-CCAF2E645FA1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{101D9C9C-23C5-40AA-A927-CCAF2E645FA1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{101D9C9C-23C5-40AA-A927-CCAF2E645FA1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B6F86E78-30C8-488F-AE8E-FFEC90FB3A13}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B6F86E78-30C8-488F-AE8E-FFEC90FB3A13}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B6F86E78-30C8-488F-AE8E-FFEC90FB3A13}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B6F86E78-30C8-488F-AE8E-FFEC90FB3A13}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
